+++ m_archive/src/router/index.js
@@ -6,6 +6,9 @@ import { TestIm,TestJo,TestMoon,TestKim,TestSeo } from '../pages/Test';
 
 const router = (
     <Route path="/">
+      {/*
+      <Route index element = {<StartPage/>}/>
+*/}
       <Route path='test'>
          <Route path='im' element={<TestIm/>}/>
           <Route path='moon' element={<TestMoon/>}/>
@@ -13,6 +16,29 @@ const router = (
           <Route path='jo' element={<TestJo/>}/>
           <Route path='seo' element={<TestSeo/>}/>
       </Route>
+      {/*
+      <Route path="login" element={<LoginPage/>}/>
+      <Route path="register" element={<RegisterPage/>}/>
+
+      <Route path="home" element={<MovieHome/>}>
+        <Route path=":id" element={<MovieDetailPage/>}/>
+      </Route>
+      <Route path="mypage" >
+        <Route index element={<ProfilePage/>}/>
+        <Route path="like" element={<LikePage/>}/>
+        <Route path="review" element={<ReviewPage/>}/>
+        <Route path="faq" element={<FAQ/>}/>
+        <Route path=":id" element={<UserPage/>}/>
+      </Route>
+      <Route path="admin">
+        <Route index element={<ManageMoviesPage/>}>
+          <Route path="addMovie" element={<AddMoviePage/>}>
+        </Route>
+        <Route path="manageReviews" element={<ManageReviewsPage/>}/>
+        <Route path="manageUsers" element={<ManageUsersPage/>}/>
+        <Route path="manageFAQs" element={<ManageFAQsPage/>}/>
+      </Route>
+      */}
     </Route>
 )
